refactor(dao): In-Memory Dao 삭제, DB Dao를 Controller에 연결
