+++ src/test/java/wooteco/subway/acceptance/AcceptanceTest.java
@@ -1,13 +1,10 @@
 package wooteco.subway.acceptance;
 
 import io.restassured.RestAssured;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
-import wooteco.subway.dao.LineDao;
-import wooteco.subway.dao.StationDao;
 
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -18,11 +15,5 @@ public class AcceptanceTest {
     @BeforeEach
     public void setUp() {
         RestAssured.port = port;
-    }
-
-    @AfterEach
-    public void rollback() {
-        StationDao.deleteAll();
-        LineDao.deleteAll();
     }
 }
