+++ src/test/java/wooteco/subway/dao/LineDaoTest.java
@@ -4,32 +4,57 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.domain.Line;
+import wooteco.subway.exception.NotFoundException;
 
-class LineDaoTest {
+@JdbcTest
+public class LineDaoTest {
 
-    @AfterEach
-    private void rollback() {
-        LineDao.deleteAll();
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    private LineDao lineDao;
+
+    @BeforeEach
+    void setup() {
+        lineDao = new LineDao(jdbcTemplate);
     }
 
     @Test
-    @DisplayName("노선 이름이 중복되면 예외가 발생한다.")
-    void save_inValidName() {
+    @DisplayName("이름이 일치하는 노선 조회한다.")
+    void findByName() {
+        //given
+        String color = "bg-red-600";
+        String name = "7호선";
+        final Line line = new Line(name, color);
+        lineDao.save(line);
+
+        //when
+        Line foundLine = lineDao.findByName(name);
+
+        //then
+        assertThat(foundLine.getName()).isEqualTo(name);
+        assertThat(foundLine.getColor()).isEqualTo(color);
+    }
+
+    @Test
+    @DisplayName("노선을 저장하면 저장된 노선 정보를 반환한다.")
+    void save() {
         // given
         final Line line = new Line("7호선", "bg-red-600");
-        LineDao.save(line);
 
         // when
-        final Line newLine = new Line("7호선", "bg-green-600");
+        Line savedLine = lineDao.save(line);
 
         // then
-        assertThatThrownBy(() -> LineDao.save(newLine))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("중복된 이름의 노선은 저장할 수 없습니다.");
+        assertThat(savedLine.getName()).isEqualTo(line.getName());
+        assertThat(savedLine.getColor()).isEqualTo(line.getColor());
     }
 
     @Test
@@ -37,15 +62,15 @@ class LineDaoTest {
     void findAll() {
         // given
         final Line line7 = new Line("7호선", "bg-red-600");
-        LineDao.save(line7);
+        lineDao.save(line7);
 
         final String line5Name = "5호선";
         final String line5Color = "bg-green-600";
         final Line line5 = new Line(line5Name, line5Color);
-        LineDao.save(line5);
+        lineDao.save(line5);
 
         // when
-        final List<Line> lines = LineDao.findAll();
+        final List<Line> lines = lineDao.findAll();
 
         // then
         assertThat(lines.size()).isEqualTo(2);
@@ -57,10 +82,10 @@ class LineDaoTest {
     void findById() {
         // given
         final Line line7 = new Line("7호선", "bg-red-600");
-        Line persistLine = LineDao.save(line7);
+        Line persistLine = lineDao.save(line7);
 
         // when
-        Line actual = LineDao.findById(persistLine.getId());
+        Line actual = lineDao.findById(persistLine.getId());
 
         // then
         assertThat(actual).isEqualTo(persistLine);
@@ -69,8 +94,8 @@ class LineDaoTest {
     @Test
     @DisplayName("존재하지 않은 id로 노선을 조회하면 예외가 발생한다.")
     void findById_invalidId() {
-        assertThatThrownBy(() -> LineDao.findById(1L))
-                .isInstanceOf(IllegalArgumentException.class)
+        assertThatThrownBy(() -> lineDao.findById(1L))
+                .isInstanceOf(NotFoundException.class)
                 .hasMessage("해당 ID에 맞는 노선을 찾지 못했습니다.");
     }
 
@@ -78,11 +103,11 @@ class LineDaoTest {
     @DisplayName("id를 통해 해당하는 노선을 업데이트한다.")
     void updateById() {
         // given
-        final Line persistLine = LineDao.save(new Line("7호선", "bg-red-600"));
+        final Line persistLine = lineDao.save(new Line("7호선", "bg-red-600"));
 
         // when
         final Line line = new Line("5호선", "bg-green-600");
-        final Line updatedLine = LineDao.updateById(persistLine.getId(), line);
+        final Line updatedLine = lineDao.updateById(persistLine.getId(), line);
 
         // then
         assertThat(updatedLine).isEqualTo(line);
@@ -92,14 +117,14 @@ class LineDaoTest {
     @DisplayName("id를 통해 해당하는 노선을 삭제한다.")
     void deleteById() {
         // given
-        final Line persistLine = LineDao.save(new Line("7호선", "bg-red-600"));
+        final Line persistLine = lineDao.save(new Line("7호선", "bg-red-600"));
 
         // when
         Long id = persistLine.getId();
-        Integer affectedRows = LineDao.deleteById(id);
+        Integer affectedRows = lineDao.deleteById(id);
 
         // then
         assertThat(affectedRows).isOne();
-        assertThat(LineDao.findAll()).hasSize(0);
+        assertThat(lineDao.findAll()).hasSize(0);
     }
 }
