+++ src/test/java/wooteco/subway/dao/StationDaoTest.java
@@ -1,26 +1,67 @@
 package wooteco.subway.dao;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.List;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 import wooteco.subway.domain.Station;
 
+@JdbcTest
 class StationDaoTest {
 
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
+    private StationDao stationDao;
+
+    @BeforeEach
+    void setUp() {
+        stationDao = new StationDao(jdbcTemplate);
+    }
+
     @Test
-    @DisplayName("역 이름이 중복되면 예외가 발생한다.")
-    void save_inValidName() {
+    @DisplayName("역을 저장하면 저장된 역 정보를 반환한다.")
+    void save() {
         // given
-        final Station station = new Station("선릉");
-        StationDao.save(station);
+        final String name = "선릉";
+        final Station station = new Station(name);
 
         // when
-        final Station newStation = new Station("선릉");
+        final Station savedStation = stationDao.save(station);
 
         // then
-        assertThatThrownBy(() -> StationDao.save(newStation))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("중복된 이름의 역은 저장할 수 없습니다.");
+        assertThat(savedStation.getName()).isEqualTo(name);
+    }
+
+    @Test
+    @DisplayName("모든 역 조회하기")
+    void findAll() {
+        // given
+        stationDao.save(new Station("선릉"));
+        stationDao.save(new Station("노원"));
+
+        // when
+        List<Station> stations = stationDao.findAll();
+
+        // then
+        assertThat(stations).hasSize(2);
+    }
+
+    @Test
+    @DisplayName("id에 해당하는 역 삭제하기")
+    void deleteById() {
+        // given
+        Station station = stationDao.save(new Station("선릉"));
+
+        // when
+        Integer affectedRows = stationDao.deleteById(station.getId());
+
+        // then
+        assertThat(affectedRows).isOne();
     }
 }
